refactor: create store folder and moved api into it
