--- src/env.js
@@ -1,11 +1,0 @@
-export const CONTENTFUL_SPACE_ID = import.meta.env.VITE_CONTENTFUL_SPACE_ID;
-export const CONTENTFUL_ACCESS_TOKEN = import.meta.env.VITE_CONTENTFUL_ACCESS_TOKEN;
-
-export const GITHUB_API_TOKEN = import.meta.env.VITE_GITHUB_API_TOKEN;
-export const GITHUB_REPOS = import.meta.env.VITE_GITHUB_REPOS;
-
-export const KOALA_API_BASE = import.meta.env.VITE_KOALA_API_BASE;
-export const LOGO = import.meta.env.VITE_LOGO;
-
-export const NEXT_INTERVAL = import.meta.env.VITE_NEXT_INTERVAL;
-export const NEXT_INTERVAL_COMMITS = import.meta.env.VITE_NEXT_INTERVAL_COMMITS;
